backend/requirements.txt
autopep8==1.6.0
dj-rest-auth==2.2.4
django-allauth==0.50.0
django-anymail==8.5
django-cors-headers==3.10.1
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-haystack==3.2.1
django-phonenumber-field==6.3.0
django-rest-auth==0.9.5
Django==3.2.12
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
gevent==21.12.0
google-auth==2.16.1
greenlet==1.1.2
gunicorn==20.1.0
ipython==8.8.0
phonenumbers==8.12.51
Pillow==9.2.0
psycopg2-binary==2.8.6
redis==4.0.2
Werkzeug==2.0.3